requirements.txt
gradio==4.16.0
scikit-learn==1.2.2
numpy==1.24.4
pandas==2.0.3
joblib==1.3.2
google-generativeai
pydantic
fastapi
uvicorn
gunicorn 